- Fix move logic.

+++ src/main/java/com/compiler/Main.java
@@ -109,10 +109,10 @@ public class Main {
     rootModule.functions.add(function0);
 
     // Locals
-    var localVar0 = symbolFactory.CreateSymbolLocalVariable(function0, "temp0", DtInteger.i16);
+    var localVar0 = symbolFactory.CreateSymbolLocalVariable(function0, "temp0", DtInteger.i32);
     function0.locals.add(localVar0);
 
-    var localVar1 = symbolFactory.CreateSymbolLocalVariable(function0, "temp1", DtInteger.i32);
+    var localVar1 = symbolFactory.CreateSymbolLocalVariable(function0, "temp1", DtInteger.i64);
     function0.locals.add(localVar1);
 
     // Code
